33 - Autogenerate the schema on Graph initialisation.
