+++ client/tinychain/__init__.py
@@ -1,23 +1,54 @@
 import tinychain.app
 import tinychain.error
 import tinychain.graph
-import tinychain.host
 import tinychain.ml
 import tinychain.ml.service
 import tinychain.scalar
-
-from tinychain.collection import Column
-from tinychain.collection import btree, table, tensor
-from tinychain.context import print_json, to_json, Context
+from tinychain.registry import Registry
+from tinychain.collection import Column, btree, table, tensor
+from tinychain.context import Context, print_json, to_json
 from tinychain.decorators import closure, delete, differentiable, get, post, put
+from tinychain.generic import Map, Tuple
 from tinychain.graph import Edge, Graph
+from tinychain.interface import Equality, Interface, Order
 from tinychain.math import linalg
 from tinychain.reflect import Meta
-from tinychain.generic import Map, Tuple
-from tinychain.interface import Interface, Equality, Order
-from tinychain.scalar.ref import After, Case, If, While  # TODO: replace with helper methods
-from tinychain.scalar.ref import form_of, get_ref, is_literal, is_conditional, is_none, is_op, is_op_ref, is_ref, same_as
+from tinychain.scalar.number import (
+    C32,
+    C64,
+    F32,
+    F64,
+    I16,
+    I32,
+    I64,
+    U8,
+    U16,
+    U32,
+    U64,
+    Bool,
+    Complex,
+    Float,
+    Int,
+    Number,
+    UInt,
+)
+from tinychain.scalar.ref import (
+    After,
+    Case,
+    If,
+    While,
+    form_of,
+    get_ref,
+    is_conditional,
+    is_literal,
+    is_none,
+    is_op,
+    is_op_ref,
+    is_ref,
+    same_as,
+)  # TODO: replace with helper methods
 from tinychain.scalar.value import EmailAddress, Id, String, Value
-from tinychain.scalar.number import Number, Bool, Complex, C32, C64, Float, F32, F64, Int, I16, I32, I64, UInt, U8, U16, U32, U64
 from tinychain.state import Class, Instance, State, Stream
 from tinychain.uri import URI
+
+import tinychain.host
