+++ client/tinychain/app.py
@@ -1,5 +1,4 @@
 """A hosted :class:`App` or :class:`Library`."""
-
 import inspect
 import logging
 import typing
@@ -398,34 +397,3 @@ def _is_mutable(state):
         return False
 
     return True
-
-
-def create_schema(modelclass: typing.Type[Model]) -> Schema:
-    """Create a table schema for the given model. A key for the table is auto
-    generated using the `class_name` function, then suffixed with '_id'. Each
-    attribute of the model will be considered as a column if it is of type
-    Column or Model.
-    """
-    values = []
-    indices = []
-    base_attributes = set()
-
-    for b in modelclass.__bases__:
-        base_attributes |= set(dir(b))
-
-    for f in base_attributes ^ set(dir(modelclass)):
-        attr = getattr(modelclass, f)
-        if isinstance(attr, Column):
-            values.append(attr)
-        else:
-            try:
-                assert issubclass(attr, Model)
-                values.append(*attr.key())
-                indices.append((attr.class_name(), [attr.key()[0].name]))
-            except (TypeError, AssertionError):
-                continue
-
-    schema = Schema(modelclass.key(), values)
-    for i in indices:
-        schema.create_index(*i)
-    return schema
