+++ tests/requirements.txt
@@ -1,0 +1,21 @@
+black==22.3.0
+certifi==2022.5.18.1
+charset-normalizer==2.0.12
+click==8.1.3
+docker==5.0.3
+docopt==0.6.2
+idna==3.3
+isort==5.10.1
+mypy-extensions==0.4.3
+num2words==0.5.10
+numpy==1.22.4
+pathspec==0.9.0
+platformdirs==2.5.2
+requests==2.27.1
+tomli==2.0.1
+torch==1.11.0
+typing_extensions==4.2.0
+urllib3==1.26.9
+websocket-client==1.3.2
+-e .
+-e tests/
