+++ tests/tctest/__main__.py
@@ -1,6 +1,5 @@
 import unittest
 
-from .apps import *
 from .client import *
 from .host import *
 from .unit import *
