+++ tests/tctest/client/__main__.py
@@ -5,5 +5,4 @@ from .test_operators import *
 from .test_table import *
 from .test_tensor import *
 
-
 unittest.main()
