+++ tests/tctest/unit/__main__.py
@@ -2,8 +2,10 @@ import logging
 import sys
 import unittest
 
+from .collection.test_graph import *
+from .collection.test_table import *
 from .test_app import *
-from .test_graph import *
+from .test_registry import *
 
 logging.basicConfig(level=logging.DEBUG, stream=sys.stdout)
